Fix MSYS2 build error
Fix use of non-standard integer type "uint". This is not
available in gcc under MSYS2.

../hw/i2c/esp32_i2c.c: In function 'esp32_i2c_do_transaction':
../hw/i2c/esp32_i2c.c:206:22: error: unknown type name 'uint'; did you mean 'u_int'?
  206 |                 for (uint nbytes = 0; nbytes < length; ++nbytes) {
      |                      ^~~~
      |                      u_int

+++ hw/i2c/esp32_i2c.c
@@ -203,7 +203,7 @@ static void esp32_i2c_do_transaction(Esp32I2CState * s)
                     s->int_raw_reg = FIELD_DP32(s->int_raw_reg, I2C_INT_RAW, ACK_ERR, 0);
                     length -= 1;
                 }
-                for (uint nbytes = 0; nbytes < length; ++nbytes) {
+                for (size_t nbytes = 0; nbytes < length; ++nbytes) {
                     uint8_t data = fifo8_pop(&s->tx_fifo);
                     i2c_send(s->bus, data);
                 }
